pyproject.toml
[project]
name = "ai-image-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.5.0",
    "modal>=1.3.2",
    "pytorch-ignite>=0.5.3",
    "tqdm>=4.67.3",
]
